requirements.txt
langchain>=0.1.0
langchain-community>=0.0.2
requests>=2.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
unstructured>=0.10.0